--- backend/src/main/java/com/stampcrush/backend/entity/CafePolicy.java
@@ -1,28 +1,0 @@
-package com.stampcrush.backend.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-
-import static jakarta.persistence.FetchType.LAZY;
-import static jakarta.persistence.GenerationType.IDENTITY;
-
-@Getter
-@Entity
-public class CafePolicy extends BaseDate {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    private Integer maxStampCount;
-
-    private String reward;
-
-    private Integer expirePeriod;
-
-    private Boolean deleted;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "cafe_id")
-    private Cafe cafe;
-}
