+++ backend/src/main/java/com/stampcrush/backend/entity/cafe/CafeCouponDesign.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.cafe;
 
+import com.stampcrush.backend.entity.baseentity.BaseDate;
 import jakarta.persistence.*;
 import lombok.Getter;
