+++ backend/src/main/java/com/stampcrush/backend/entity/cafe/CafePolicy.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.cafe;
 
+import com.stampcrush.backend.entity.baseentity.BaseDate;
 import jakarta.persistence.*;
 import lombok.Getter;
 
@@ -8,19 +9,19 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Getter
 @Entity
-public class Reward extends BaseDate {
+public class CafePolicy extends BaseDate {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
-    private String name;
+    private Integer maxStampCount;
 
-    private Boolean used;
+    private String reward;
 
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "customer_id")
-    private Customer customer;
+    private Integer expirePeriod;
+
+    private Boolean deleted;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "cafe_id")
