+++ backend/src/main/java/com/stampcrush/backend/entity/coupon/Coupon.java
@@ -1,0 +1,40 @@
+package com.stampcrush.backend.entity.coupon;
+
+import com.stampcrush.backend.entity.user.Customer;
+import com.stampcrush.backend.entity.baseentity.BaseDate;
+import com.stampcrush.backend.entity.cafe.Cafe;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.time.LocalDate;
+
+import static jakarta.persistence.FetchType.LAZY;
+import static jakarta.persistence.GenerationType.IDENTITY;
+
+@Getter
+@Entity
+public class Coupon extends BaseDate {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    private LocalDate expiredDate;
+
+    @Enumerated(EnumType.STRING)
+    private CouponStatus status;
+
+    private Boolean deleted;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "customer_id")
+    private Customer customer;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "cafe_id")
+    private Cafe cafe;
+
+    @OneToOne(fetch = LAZY)
+    @JoinColumn(name = "coupon_design_id")
+    private CouponDesign couponDesign;
+}
