+++ backend/src/main/java/com/stampcrush/backend/entity/coupon/CouponDesign.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.coupon;
 
+import com.stampcrush.backend.entity.baseentity.BaseDate;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -9,15 +10,14 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Getter
 @Entity
-public class CouponPolicy extends BaseDate {
-
+public class CouponDesign extends BaseDate {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
-    private Integer maxStampCount;
+    private String frontImageUrl;
 
-    private String rewardName;
+    private String backImageUrl;
 
-    private Integer expiredPeriod;
+    private String stampImageUrl;
 }
