+++ backend/src/main/java/com/stampcrush/backend/entity/coupon/CouponPolicy.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.coupon;
 
+import com.stampcrush.backend.entity.baseentity.BaseDate;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -9,13 +10,15 @@ import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Getter
 @Entity
-public class Owner extends BaseDate {
+public class CouponPolicy extends BaseDate {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
-    private String name;
-    private String loginId;
-    private String encryptedPassword;
-    private String phoneNumber;
+
+    private Integer maxStampCount;
+
+    private String rewardName;
+
+    private Integer expiredPeriod;
 }
