+++ backend/src/main/java/com/stampcrush/backend/entity/coupon/CouponStatus.java
@@ -1,4 +1,4 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.coupon;
 
 public enum CouponStatus {
     USING, EXPIRED, REWARDED;
