+++ backend/src/main/java/com/stampcrush/backend/entity/reward/Reward.java
@@ -1,27 +1,25 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.reward;
 
+import com.stampcrush.backend.entity.baseentity.BaseDate;
+import com.stampcrush.backend.entity.cafe.Cafe;
+import com.stampcrush.backend.entity.user.Customer;
 import jakarta.persistence.*;
 import lombok.Getter;
-
-import java.time.LocalDate;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Getter
 @Entity
-public class Coupon extends BaseDate {
+public class Reward extends BaseDate {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
-    private LocalDate expiredDate;
+    private String name;
 
-    @Enumerated(EnumType.STRING)
-    private CouponStatus status;
-
-    private Boolean deleted;
+    private Boolean used;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "customer_id")
@@ -30,8 +28,4 @@ public class Coupon extends BaseDate {
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "cafe_id")
     private Cafe cafe;
-
-    @OneToOne(fetch = LAZY)
-    @JoinColumn(name = "coupon_design_id")
-    private CouponDesign couponDesign;
 }
