+++ backend/src/main/java/com/stampcrush/backend/entity/user/RegisterCustomer.java
@@ -1,4 +1,4 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.user;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
