+++ backend/src/main/java/com/stampcrush/backend/entity/user/TemporaryCustomer.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.entity;
+package com.stampcrush.backend.entity.user;
 
+import com.stampcrush.backend.entity.user.Customer;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
